+++ tardis/montecarlo/transport/opacities.py
@@ -1,10 +1,10 @@
 from numba import njit
 
-from tardis.montecarlo.montecarlo_numba import (
+from tardis.montecarlo.transport import (
     njit_dict_no_parallel,
 )
 
-from tardis.montecarlo.montecarlo_numba.numba_config import (
+from tardis.montecarlo.transport.numba_config import (
     SIGMA_THOMSON,
 )
