+++ tardis/montecarlo/transport/r_packet_transport.py
@@ -2,18 +2,18 @@ import numpy as np
 from numba import njit
 
 from tardis.montecarlo import montecarlo_configuration
-from tardis.montecarlo.montecarlo_numba import njit_dict_no_parallel
-from tardis.montecarlo.montecarlo_numba.calculate_distances import \
+from tardis.montecarlo.transport import njit_dict_no_parallel
+from tardis.montecarlo.transport.calculate_distances import \
     calculate_distance_boundary, calculate_distance_electron, \
     calculate_distance_line
-from tardis.montecarlo.montecarlo_numba.estimators import \
+from tardis.montecarlo.transport.estimators import \
     update_line_estimators, set_estimators, set_estimators_full_relativity
-from tardis.montecarlo.montecarlo_numba.frame_transformations import \
+from tardis.montecarlo.transport.frame_transformations import \
     get_doppler_factor
-from tardis.montecarlo.montecarlo_numba.numba_config import \
+from tardis.montecarlo.transport.numba_config import \
     ENABLE_FULL_RELATIVITY
-from tardis.montecarlo.montecarlo_numba.opacities import calculate_tau_electron
-from tardis.montecarlo.montecarlo_numba.r_packet import InteractionType, \
+from tardis.montecarlo.transport.opacities import calculate_tau_electron
+from tardis.montecarlo.transport.r_packet import InteractionType, \
     PacketStatus
 
 
@@ -178,7 +178,7 @@ def move_r_packet(r_packet, distance, time_explosion, numba_estimator):
         r_packet objects
     time_explosion : float
         time since explosion in s
-    numba_estimator : tardis.montecarlo.montecarlo_numba.numba_interface.NumbaEstimator
+    numba_estimator : tardis.montecarlo.transport.numba_interface.NumbaEstimator
         Estimators object
     distance : float
         distance in cm
