+++ tardis/montecarlo/transport/utils.py
@@ -1,6 +1,6 @@
 import numpy as np
 from numba import njit
-from tardis.montecarlo.montecarlo_numba import (
+from tardis.montecarlo.transport import (
     njit_dict_no_parallel,
 )
